Also use SVLEN for <DUP> and <CNV> alleles
After discussion and more careful reading of the VCF specification,
add <DUP>, <CNV> and <CNV: to the set of symbolic alleles that
should take account of SVLEN in the rlen computation.  The last
one is to catch <CNV:TR>, but is left general in case other
CNV variants appear, on the assumption that they will use SVLEN
in the same way as the existing tags.