Check for x86intrin.h header and ensure it's used consistently
Add checks for `x86intrin.h` for both configure and make-only
builds, and use it to gate x86intrin.h inclusion.

Update `simd.c` to use it in preference to `immintrin.h`.
While the latter is the one preferred by Intel, and possibly
more portable than `x86intrin.h` (which was introduced by gcc
and clang), htscodecs consistently uses `x86intrin.h` so it's
easier to standardise on that one.  The intel compiler now supports
both, so the portability issue is less of a problem than it used
to be.