Add a way of attaching private data to a BGZF file handle
The motivation for this is to enable passing of a pointer to
a bcf_hdr_t structure to bcf_readrec(), which currently does
not get one.  It does always get a pointer for the BGZF handle,
so a header struct could be passed in via that if it can be
stored somewhere.

To enable this while not changing the bgzf API or ABI, extra
fields are added to the opaque bgzf_cache_t field.  The BGZF_CACHE
macro that could be use to disable addition of the cache feature
removed as it was always turned on anyway.  The cache struct now
has to be created for files open for write, although the cache
part is not used.  The hash type used by the cache is renamed from
"cache" to "bgzf_cache" to improve its name-spacing.

The interfaces to add, get, and remove private data are put in
a new bgzf_internal.h header.  The bgzf_cache_t struct definition
is also moved there so that the get function can be inlined for
faster access to the private data field.

The bgzf_cache_t definition is rewritten slightly so that it's
not necessary to invoke KHASH_MAP_INIT_INT64() before it in the
header file, as doing that would require struct cache_t to be
moved from bgzf.c to the new header as well.  Instead, typedef
kh_bgzf_cache_t is used in place of khash(bgzf_cache), and
unsigned int instead of khint_t.