For the VCF rlen calculation, only use SVLEN for DEL, DUP and CNV symbolic alleles (#1942)
* For rlen calculation, only use SVLEN for DEL symbolic alleles

PR #1897 updated the VCF rlen calculation in line with VCF 4.4,
including taking account of INFO SVLEN for non-INS alleles.  This
was not correct though as it included non-symbolic alleles if they
(incorrectly) had a non-missing SVLEN; and other types of symbolic
allele that do not affect the length with respect to the reference.
In fact, only deletions can have an affect on rlen, so this changes
the calculation to only use SVLEN for <DEL> alleles.

A bug is also fixed on big-endian platforms where INFO and FORMAT
values were being accessed incorrectly.  These are always stored
in little-endian order, so need to be converted to native ordering
when read.  The tests are updated to include values greater than
128 to ensure that this works correctly.

* Also use SVLEN for <DUP> and <CNV> alleles

After discussion and more careful reading of the VCF specification,
add <DUP>, <CNV> and <CNV: to the set of symbolic alleles that
should take account of SVLEN in the rlen computation.  The last
one is to catch <CNV:TR>, but is left general in case other
CNV variants appear, on the assumption that they will use SVLEN
in the same way as the existing tags.

* Use SVLEN for <INV> alleles plus CNV, DEL, DUP and INV subtypes

As INV is defined as "inversion of the reference sequence", it
needs to be added to the list of symbolic ALT alleles where
the reference length is stored in SVLEN.  The code to check which
symbolic ALT types do this is also extended to support subtypes
for all of CNV, DEL, DUP and INV; and moved to hts_internal.h
so it doesn't need to be repeated in both vcf.c and tbx.c.

* Minor get_rlen() speed-ups

Delay looking up ids for LEN and SVLEN until it's known that they
are needed.  Also make a quick test to see if alleles are symbolic
before trying to work out what sort they are.