Improve CRAM TLEN calculation.
Specifically when we have the same start coordinate, we need to be
checking the end coordinate too as the stated rule is leftmost-this,
rightmost-that and we can't just resolve ties on leftmost without
first checking if rightmost is unambiguous.

This now matches the specification, but the spec should be clarified a
bit better too with regards to ties.  It also lends itself very much
to the incorrect logic we had before as it's pseudocode implying a
strict order of instructions rather than assessing both ends
simultaneously.  It would be better written in the pseudocode with "if
leftmost do-this, else do-that" style.  (Incoming as a separate PR)

Also cope with multiple sets of R1 and R2 reads appearing in a template.
If we have multiple pairs of internally consistent secondary alignments
(or multiple primaries, but that's invalid anyway) then we don't want to
form a single PNEXT/RNEXT chain for all secondaries as this may not be
what the aligner produced, even though it's what the SAM specification
implies.

We track the number of times READ1 and READ2 have occurred and once we
observe a second copy we switch to detached mode - storing verbatim.

This isn't perfect as we could have interleaved sets, such as:

    1---->.........<------2
            2--->..............<---1

If this is position sorted then we'd end up processing the 1> and 2>
together before realising we have a duplicated <2 and <1.  However
fixing this requires a second pass through the data, as would doing
round-trip tests of encode, decode, validate.  Our encoder as written
now is a "process_one_read" loop so unless we want to completely
rewrite it then all we can do is cache what's happened before to
detect when we switch into detached mode.

However the fix is an improvement in many problematic cases.
On real-world data the different is likely to be minimal or
non-existant.