Fix possible memory leak on successful match in kmemmem()
Add missing free() for the case where kmemmem() found a match
and the helper data in `prep` was not either passed in, or
passed back to the caller via `_*prep`.

Also make the logic around setting `prep` and potentially passing
it back easier to understand.

Credit to OSS_Fuzz
Fixes oss-fuzz issue 444490034