Fix embed_ref=2 on SEQ * and MD:Z tag.
The combination of no sequence and MD:Z with embed_ref=2 caused the
slice extents to be miscalculated.  This caused decode errors due to
validation of slice extents on decode.  It wouldn't matter as SEQ *
means there's nothing to decode anyway, but the CRAM container / slice
headers are still incorrect and it's valid to complain.

We now spot this issue and still compute the reference end using
CIGAR.  MD without CIGAR isn't supported, however htslib has had a
long standing disagreement with the SAM spec by treating CIGAR "*" as
unmapped.

Fixes samtools/samtools#2277