Don't drop in-flight decode jobs when seeking in cram_next_slice()
cram_next_slice() with a region set tries to skip over containers
with contents that don't overlap the region of interest.  For
files with a mixture of read lengths, it's possible that some
decode jobs have been queued before a later container is skipped.
This makes calling cram_seek() hazardous as it drops any in-flight
jobs before updating the file position.  Instead, call hseek()
directly so already-queued decode jobs are retained.