S3 Code Rewrite (PR #1958)
Add code to read from S3 sources in the same style as the current writing code.
Merge the s3_write.c code into hfile_s3.c to put all the S3 code in the same
file.

This should simplify the S3 code and give us a little more flexibility in how
we use it.