Fix htslib.map update
At some point `nm --with-symbol-versions` stopped appending
`@@Base` to unversioned symbols.  This broke the awk script
that looked for symbols to add to the `htslib.map` file.  Fix
by changing the match criteria to accept symbol names with no
version appended as well as `@@Base`, and make minor adjustments
to how the semicolon is appended to the name in the output.

Also add `--defined-only` to the `nm` command line to get rid
of imported functions as they don't follow the same output
format as for the symbols we want (the first column is blank).
Removing them eliminates the chance that these lines could
be incorrectly included.