Simplify kh_grow_to_fit table size check
For some reason, early versions of clang incorrectly optimised a
check to ensure that the number of buckets in a resized hash table
will still fit in a khint_t, with the result that most of the
function was replaced by "return -2".  Fix by putting the limit
value in a separate variable, so the comparison part no longer
needs to do any type conversion.