Add overflow check in bgzf_index_load_hfile()
This function reads an item count from the input file and uses
it to allocate an array of file offsets.  It's possible for
this to overflow, resulting in an attempt to malloc(0) and,
if that returns a valid pointer, an attempt to zero 16 bytes
in a zero-length allocation.

Fix by checking that the allocation will not overflow.

Thanks to Harrison Green for reporting this.

(cherry picked from commit 6dd0d7d0e9e7e2e173a28969e624db8bc8bb5828)